refactor(waveform-seek): H3.1 — extract helpers + constants + types
